Change the int value to float and make it available to handle the decimal number

+++ assets/scripts/app.js
@@ -1,33 +1,43 @@
+// this is the current result
 let currentResult = 0; 
 
+// parse the user input to float
 let parseUserInput = function (){
     const userInputString= userInput.value;
-    const userInputNumber = parseInt(userInputString);
+    const userInputNumber = parseFloat(userInputString);
     return userInputNumber;
 }
 
+// this function responsible to create calculation log
 let createCalculationLog = function (operator, numberInputed){
     return `${currentResult} ${operator} ${numberInputed}`
 }
 
+// add buton clicked
 function onClickedAddBtn(){
     addCalculation = createCalculationLog("+", parseUserInput());
     currentResult += parseUserInput();
     outputResult(currentResult, addCalculation);
     console.log(userInput.value);
 }
+
+// subtract button clicked
 function onClickedSubtractBtn(){
     subtractCalculation = createCalculationLog("-", parseUserInput());
     currentResult -= parseUserInput();
     outputResult(currentResult, subtractCalculation);
     console.log(userInput.value);
 }
+
+// multiply button clicked
 function onClickedMultiplyBtn(){
     multiplyCalculation = createCalculationLog("*", parseUserInput());
     currentResult *= parseUserInput();
     outputResult(currentResult, multiplyCalculation);
     console.log(userInput.value);
 }
+
+// Divide button clicked
 function onClickedDivideBtn(){
     divideCalculation = createCalculationLog("/", parseUserInput());
     currentResult /= parseUserInput();
@@ -35,6 +45,7 @@ function onClickedDivideBtn(){
     console.log(userInput.value);
 }
 
+// Event for each button
 addBtn.addEventListener('click', onClickedAddBtn);
 subtractBtn.addEventListener('click', onClickedSubtractBtn);
 multiplyBtn.addEventListener('click', onClickedMultiplyBtn);
